Moving some code to the base class
